requirements.txt
aiofiles==24.1.0
aiohappyeyeballs==2.6.1
aiohttp==3.12.15
aiolimiter==1.2.1
aiosignal==1.4.0
altair==5.5.0
annotated-types==0.7.0
anyio==4.10.0
anytree==2.13.0
ascii_colors==0.11.4
asttokens==2.4.1
async-timeout==5.0.1
attrs==25.3.0
autograd==1.8.0
azure-common==1.1.28
azure-core==1.35.0
azure-cosmos==4.9.0
azure-identity==1.24.0
azure-search-documents==11.5.3
azure-storage-blob==12.26.0
backports-datetime-fromisoformat==2.0.3
beartype==0.18.5
blinker==1.9.0
blis==1.2.1
cachetools==6.2.0
catalogue==2.0.10
certifi==2025.8.3
cffi==1.17.1
charset-normalizer==3.4.3
click==8.2.1
cloudpathlib==0.22.0
confection==0.1.5
configparser==7.2.0
contourpy==1.3.2
cryptography==45.0.7
cycler==0.12.1
cymem==2.0.11
deprecation==2.1.0
devtools==0.12.2
distro==1.9.0
dotenv==0.9.9
environs==14.3.0
exceptiongroup==1.3.0
executing==2.2.1
fnllm==0.4.1
fonttools==4.59.2
frozenlist==1.7.0
future==1.0.0
gensim==4.3.3
gitdb==4.0.12
GitPython==3.1.45
graphrag==2.5.0
graspologic==3.4.1
graspologic-native==1.2.5
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
hyppo==0.4.0
idna==3.10
isodate==0.7.2
Jinja2==3.1.6
jiter==0.10.0
joblib==1.5.2
json_repair==0.50.0
jsonschema==4.25.1
jsonschema-specifications==2025.4.1
kiwisolver==1.4.9
lancedb==0.24.3
langcodes==3.5.0
language_data==1.3.0
lightrag-hku==1.4.7
llvmlite==0.44.0
marisa-trie==1.3.1
markdown-it-py==4.0.0
MarkupSafe==3.0.2
marshmallow==4.0.1
matplotlib==3.10.6
mdurl==0.1.2
msal==1.33.0
msal-extensions==1.3.1
multidict==6.6.4
murmurhash==1.0.13
nano-vectordb==0.0.4.3
narwhals==2.3.0
networkx==3.4.2
nltk==3.9.1
numba==0.61.2
numpy==1.26.4
openai==1.102.0
overrides==7.7.0
packaging==25.0
pandas==2.3.2
patsy==1.0.1
pillow==11.3.0
pipmaster==1.0.4
POT==0.9.5
preshed==3.0.10
propcache==0.3.2
protobuf==6.32.0
pyarrow==21.0.0
pycparser==2.22
pydantic==2.11.7
pydantic_core==2.33.2
pydeck==0.9.1
Pygments==2.19.2
PyJWT==2.10.1
pynndescent==0.5.13
pyparsing==3.2.3
pypinyin==0.55.0
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
pytz==2025.2
PyYAML==6.0.2
rdflib==7.1.4
referencing==0.36.2
regex==2025.8.29
requests==2.32.5
rich==14.1.0
rpds-py==0.27.1
scikit-learn==1.7.1
scipy==1.12.0
seaborn==0.13.2
shellingham==1.5.4
six==1.17.0
smart_open==7.3.0.post1
smmap==5.0.2
sniffio==1.3.1
spacy==3.8.7
spacy-legacy==3.0.12
spacy-loggers==1.0.5
srsly==2.5.1
statsmodels==0.14.5
streamlit==1.49.1
streamlit-agraph==0.0.45
tenacity==9.1.2
textblob==0.19.0
thinc==8.3.4
threadpoolctl==3.6.0
tiktoken==0.11.0
toml==0.10.2
tornado==6.5.2
tqdm==4.67.1
typer==0.17.3
typing-inspection==0.4.1
typing_extensions==4.15.0
tzdata==2025.2
umap-learn==0.5.9.post2
urllib3==2.5.0
wasabi==1.1.3
watchdog==6.0.0
wcwidth==0.2.13
weasel==0.4.1
wrapt==1.17.3
xlsxwriter==3.2.5
yarl==1.20.1
